# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "CancelRequest": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "const": true,
            "description": "Cancellation tears down sandboxes and stops the run; it is always confirmed explicitly.",
            "title": "Confirm",
            "type": "boolean"
          },
          "reason": {
            "default": "",
            "maxLength": 500,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "confirm"
        ],
        "title": "CancelRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The single error shape every non-2xx response uses.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "trace_id": {
            "description": "Correlates this failure with the server logs and the mission event stream. Also returned as the X-Trace-Id header.",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "trace_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "MissionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "adapter": {
            "$ref": "#/components/schemas/LanguageAdapter"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Replaying a create with the same key returns the same mission rather than creating a second one.",
            "title": "Idempotency Key"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "policy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MissionPolicy"
              }
            ]
          },
          "repository_ref": {
            "description": "Authorized repository URL or the name of an uploaded archive. Never a public target chosen at scan time.",
            "maxLength": 500,
            "minLength": 1,
            "title": "Repository Ref",
            "type": "string"
          }
        },
        "required": [
          "name",
          "repository_ref",
          "adapter"
        ],
        "title": "MissionCreateRequest",
        "type": "object"
      },
...
      "Page_MissionSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MissionSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 500,
            "minimum": 1,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MissionSummary]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerTokenAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Control API for Brahmadatta AI — an authorized, defensive Cyber-Reasoning System.\n\n**Contract status.** Frozen at D1 (issues #6, #9). Every endpoint except\n`GET /system/health` returns `501 NOT_IMPLEMENTED` with the standard error envelope;\nthe request and response schemas are final and are what the Command Center types\nagainst.\n\n**Rules the schemas enforce rather than describe.**\n\n* A verdict is derived only from the deterministic gate matrix. Model confidence is\n  recorded on `ModelProvenance` for display and cannot reach the derivation; a\n  `VerificationRecord` whose verdict does not follow from its gates cannot be\n  serialized.\n* No mission stage runs without an active authorization record.\n* Sandbox network policy is typed `\"deny\"` and has no other permitted value.\n* No inference endpoint may address a hosted third party; that is validated at\n  startup by a Django system check.\n\n**Errors.** Every non-2xx response is an `ErrorEnvelope` carrying a code from the\nfrozen vocabulary and the request's trace id, which is also returned as `X-Trace-Id`.",
    "title": "Brahmadatta AI Control API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/missions": {
      "get": {
        "operationId": "listMissions",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MissionSummary_"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "List missions",
        "tags": [
          "missions"
        ]
      },
      "post": {
        "operationId": "createMission",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionSummary"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "Create a mission draft",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}": {
      "get": {
        "operationId": "getMission",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetail"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "Mission state, stage, progress and summary counts",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/authorize": {
      "post": {
        "description": "P0-1. Until this record exists, no stage may run: `contracts.state_machine` refuses every transition past CREATED without an active record. The record is immutable and is reproduced in the evidence bundle.",
        "operationId": "authorizeMission",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizationRecord"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "Record the authorization that every later stage depends on",
        "tags": [
          "missions"
        ]
      }
    },
...
    "/api/v1/missions/{mission_id}/cancel": {
      "post": {
        "operationId": "cancelMission",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledgement"
                }
              }
            },
            "description": "Accepted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "Cancel safely and release every resource",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/events": {
      "get": {
        "description": "Server-sent events over ASGI.\n\nThe transport is real and running today; the mission events are not, because the\norchestrator that emits them does not exist yet. Rather than fabricate telemetry\n— which the product rules forbid outright — this emits the stream preamble, two\nheartbeat comments, and one terminal `contract.not_implemented` frame carrying the\nstandard error envelope, then closes.\n\nThat is enough to prove the thing most likely to break silently: nginx buffers\nproxied responses by default, and a buffered SSE stream arrives as one lump at the\nend. `X-Accel-Buffering: no` is set here as well as `proxy_buffering off` in the\nnginx config (issue #10) — belt and braces, because this failure is invisible\nuntil the demo.",
        "operationId": "streamMissionEvents",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/MissionEvent"
                }
              }
            },
            "description": "Ordered mission event stream. Each frame is `id: <sequence>`, `event: <MissionEvent.type>`, `data: <MissionEvent as JSON>`. Reconnect with `Last-Event-ID`, or replay through `/missions/{mission_id}/events/replay`."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "Ordered event stream (SSE)",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/v1/missions/{mission_id}/events/replay": {
      "get": {
        "description": "Gap recovery for the SSE stream. `sequence` is gap-free per mission, so a client that sees a jump replays from its last known value.",
        "operationId": "replayMissionEvents",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since_sequence",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Since Sequence",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MissionEvent_"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "summary": "Replay ordered events from a sequence number",
        "tags": [
          "missions"
        ]
      }
    },
...
  },
  "servers": []
}